map dialog added

+++ src/org/dyndns/fzoli/rccar/model/Point3D.java
@@ -11,7 +11,7 @@ public class Point3D implements Serializable {
     /**
      * A három irány értékei.
      */
-    public final int X, Y, Z;
+    public final double X, Y, Z;
 
     /**
      * Inicializálás és a három paraméter megadása.
@@ -19,7 +19,7 @@ public class Point3D implements Serializable {
      * @param y y irány értéke
      * @param z z irány értéke
      */
-    public Point3D(int x, int y, int z) {
+    public Point3D(double x, double y, double z) {
         X = x;
         Y = y;
         Z = z;
